+++ src/Core/Content/Product/SalesChannel/Event/ProductExtensionsLoadedEvent.php
@@ -1,0 +1,22 @@
+<?php
+declare(strict_types=1);
+
+namespace Elio\ElioDataDiscovery\Core\Content\Product\SalesChannel\Event;
+
+use Shopware\Core\System\SalesChannel\SalesChannelContext;
+use Symfony\Contracts\EventDispatcher\Event;
+
+abstract class ProductExtensionsLoadedEvent extends Event
+{
+    public function __construct(
+        private readonly SalesChannelContext $context
+    ) {}
+
+    /**
+     * @return SalesChannelContext
+     */
+    public function getContext(): SalesChannelContext
+    {
+        return $this->context;
+    }
+}