+++ src/Core/FilterRestrictions/FilterSyncService.php
@@ -223,7 +223,7 @@ class FilterSyncService
         foreach ($filterNames as $filterName) {
             $filterNameParts = explode($isSortingType ? ':' : '.', (string)$filterName);
 
-            $label = $isSortingType ? implode(' ', $filterNameParts) : ucfirst(end($filterNameParts));
+            $label = $isSortingType ? implode(' ', $filterNameParts) : $this->decodeFacetName(ucfirst(end($filterNameParts)));
 
             $createdFilters[] = [
                 'id' => Uuid::randomHex(),
@@ -242,6 +242,21 @@ class FilterSyncService
         foreach (array_chunk($createdFilters, 100) as $chunk) {
             $this->filterRepository->create($chunk, $context);
         }
+    }
+
+    /**
+     * @param string $facetName
+     * @return string
+     */
+    public function decodeFacetName(string $facetName): string
+    {
+        return preg_replace_callback(
+            '/u([0-9a-fA-F]{4})/',
+            static function ($matches) {
+                return mb_convert_encoding(pack('H*', $matches[1]), 'UTF-8', 'UCS-2BE');
+            },
+            $facetName
+        );
     }
 
     /**