+++ src/Core/Sync/Util/ProductUtil.php
@@ -325,27 +325,6 @@ class ProductUtil
      * @param array<PropertyGroupOptionEntity> $properties
      * @return array
      */
-    public static function getProductAttribute(array $properties): array
-    {
-        $attributes = [];
-        foreach ($properties as $property) {
-            $group = $property->getGroup();
-            if ($group !== null) {
-                $name = $group->getTranslation('name') ?? $group->getName();
-                $value = $property->getTranslation('name') ?? $property->getName();
-                $attributes[$name] = ValueUtil::cleanValue($value);
-            }
-        }
-
-        return $attributes;
-    }
-
-    /**
-     * Appends the product attributes
-     *
-     * @param array<PropertyGroupOptionEntity> $properties
-     * @return array
-     */
     public static function getProductProperty(array $properties): array
     {
         $attributes = [];
@@ -353,14 +332,34 @@ class ProductUtil
             $group = $property->getGroup();
             if ($group !== null) {
                 $name = $group->getTranslation('name') ?? $group->getName();
+
+                //Encode property name
+                $encodedName = self::encodePropertyName($name);
+
                 $value = $property->getTranslation('name') ?? $property->getName();
-                if (!isset($attributes[$name])) {
-                    $attributes[$name] = [];
+                if (!isset($attributes[$encodedName])) {
+                    $attributes[$encodedName] = [];
                 }
-                $attributes[$name][] = ValueUtil::cleanValue($value);
+                $attributes[$encodedName][] = ValueUtil::cleanValue($value);
             }
         }
 
         return $attributes;
+    }
+
+    public static function encodePropertyName(string $name) : string
+    {
+        $specialCharacters = ['(', ')', '[', ']', '<', '>', '`', '.', ',', ':', '=', '!', '&', '|', '$'];
+        $encodedName = '';
+
+        foreach (mb_str_split($name) as $char) {
+            if (in_array($char, $specialCharacters, true)) {
+                $encodedName .= sprintf("\\u%04x", mb_ord($char));
+            } else {
+                $encodedName .= $char;
+            }
+        }
+
+        return $encodedName;
     }
 }